workaround tailscaled bug
this bug causes tailscaled to present as ready
even though the IP has not been bound to the
interface. this will lead to sshd not being
able to start correctly.

we work around this by waiting up to 15 seconds
for the IP to come up.

this is a known bug, see
https://github.com/tailscale/tailscale/issues/11504#issue-2203591888